% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{math}
\alias{math}
\title{arithmatic
A function that provides a random sampling of basic arithmatic}
\usage{
math(a, b, c)
}
\arguments{
\item{a}{any numeric}

\item{b}{any numeric greater than 0}

\item{c}{any numeric}
}
\description{
arithmatic
A function that provides a random sampling of basic arithmatic
}
\examples{
math(4,2,4)

}
