% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow_fish.R
\name{rainbow_fish}
\alias{rainbow_fish}
\title{Rainbow Fish:
A function to give information on rainbow fish friends all over the ocean!}
\usage{
rainbow_fish(fish_name, city, fish_scales)
}
\arguments{
\item{fish_name}{the name of our fish friend}

\item{city}{the city they live in}

\item{fish_scales}{the total number of rainbow scales they have}
}
\description{
Rainbow Fish:
A function to give information on rainbow fish friends all over the ocean!
}
\examples{
rainbow_fish( "Nemo", "Coral City", 4)

}
