% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{volume}
\alias{volume}
\title{Volume:
A simple function to calculate the volume of a subastance}
\usage{
volume(L, W, H)
}
\arguments{
\item{L}{Length}

\item{W}{Width}

\item{H}{Height}
}
\description{
Volume:
A simple function to calculate the volume of a subastance
}
\examples{
volume(4, 3, 2)
}
